backend/requirements.txt
djangorestframework
django-cors-headers
markdown
django-filter
psycopg2-binary
pydevd-pycharm~=243.26053.29
requests
pydantic
pylint
pre-commit
black
flake8
isort
djangorestframework-simplejwt
django-polymorphic
django-model-utils
pillow
django-verify-email
drf-spectacular
django-enumfield
